refactor(ApplicationTest): seperate test code within ApplicationTest
ApplicationTest.java 내에 있던 RacingCar 클래스, Utils 클래스에 대한
테스트 코드를 각각 RacingCarTest.java, UtilsTest.java로 분리.
