+++ src/test/java/racingcar/ApplicationTest.java
@@ -4,11 +4,8 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberI
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static racingcar.Utils.enterCarList;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
-import java.io.ByteArrayInputStream;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class ApplicationTest extends NsTest {
@@ -32,35 +29,6 @@ class ApplicationTest extends NsTest {
                 assertThatThrownBy(() -> runException("pobi,javaji", "1"))
                         .isInstanceOf(IllegalArgumentException.class)
         );
-    }
-
-    @Test
-    void RacingCar_construct_test() {
-        RacingCar car = new RacingCar("test");
-        car.rollDice();
-        car.rollDice();
-        car.rollDice();
-        car.rollDice();
-        car.rollDice();
-        assertThat(car.getCount()).isBetween(0, 5);
-    }
-
-    @Test
-    void enterCarList_test() {
-        String input = "pobi, woni";
-        ByteArrayInputStream in = new ByteArrayInputStream(input.getBytes());
-        System.setIn(in);
-
-        assertThat(enterCarList())
-                .contains("pobi")
-                .contains("woni");
-    }
-
-    @DisplayName("자동차의 이름이 1자에서 5자 사이가 아니면 예외를 발생한다.")
-    @Test
-    void RacingCar_exception_test() {
-        assertThatThrownBy(() -> new RacingCar("invalid_name"))
-                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Override
